+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New interface `WithLifeCycle`. Change Port API to adhere to it.
 
 ### Changed
+- Separate StateContainer and MutableStateContainer
 - `DeviceState` renamed to `ValueState`
 - Update logic of `DeviceState` to properly address subscriptions.
 - `getProperty` renamed to `getCachedProperty` for `CachingDevice`
@@ -24,7 +25,7 @@
 - `DeviceHub` now works with `Name` instead of `NameToken`. Tree-like structure is made using `Path`. Device messages no longer have access to sub-devices.
 - Add some utility methods to ports. Synchronous port response could be now consumed as `Source`.
 - `DeviceLifecycleState` is replaced by `LifecycleState`.
-- Time is now mandatory first field of all device messages
+- Time is now the mandatory first field of all device messages
 
 
 ### Deprecated